refactor: вернул timetablelist в entities
