+++ src/entities/clinic/lib/getTimetable.ts
